fixed bug in completeQuest

+++ src/main/java/com/gianca1994/aowebbackend/resources/quest/QuestService.java
@@ -160,7 +160,9 @@ public class QuestService {
 
         user.userLevelUp();
 
-        userQuestRepository.delete(userQuest);
+        if (userQuest.getId() != null) userQuestRepository.delete(userQuest);
+
+        user.getUserQuests().remove(userQuest);
         userRepository.save(user);
     }
